Added authorization check in Users controller.

+++ MedicalResearch/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using UserManaging.API.Utilities;
 using UserManaging.Domain.Entities.Users;
 using UserManaging.Domain.Interfaces;
 
@@ -18,19 +19,25 @@ namespace UserManaging.API.Controllers
         [HttpPost("Create")]
         public async Task<IActionResult> CreateUserAsync(User user, CancellationToken cancellationToken)
         {
-            if(await _userService.CreateAsync(user, cancellationToken))
-                return Ok(user);
-
-            return BadRequest();
+            if (await IsCurrentUserAdminAsync())
+            {
+                if (await _userService.CreateAsync(user, cancellationToken))
+                    return Ok(user);
+            }
+            
+            return StatusCode(StatusCodes.Status403Forbidden);
         }
 
         [HttpDelete("DeleteUser")]
         public async Task<IActionResult> DeleteAsync(string userEmail, CancellationToken cancellationToken)
         {
-            if (await _userService.DeleteAsync(userEmail, cancellationToken))
-                return Ok();
+            if (await IsCurrentUserAdminAsync())
+            {
+                if (await _userService.DeleteAsync(userEmail, cancellationToken))
+                    return Ok();
+            }
 
-            return BadRequest();
+            return StatusCode(StatusCodes.Status403Forbidden);
         }
 
         [HttpGet("FindByEmail")]
@@ -47,18 +54,27 @@ namespace UserManaging.API.Controllers
         [HttpPut("Update")]
         public async Task<IActionResult> UpdateUserAsync(UserDTO user, CancellationToken cancellationToken)
         {
-            if (await _userService.UpdateAsync(user, cancellationToken))
-                return Ok();
-
-            return BadRequest();
+            if (await IsCurrentUserAdminAsync())
+            {
+                if (await _userService.UpdateAsync(user, cancellationToken))
+                    return Ok();
+            }
+            
+            return StatusCode(StatusCodes.Status403Forbidden);
         }
 
         [HttpPatch("SetUserName")]
         public async Task<IActionResult> SetUserNameAsync(UserDTO user, string userName, CancellationToken cancellationToken)
         {
-            await _userService.SetUserNameAsync(user, userName, cancellationToken);
+            var currentUser = await GetCurrentUserAsync();
 
-            return Ok();
+            if (currentUser.Email.Equals(user.Email) || await IsCurrentUserAdminAsync())
+            {
+                await _userService.SetUserNameAsync(user, userName, cancellationToken);
+                return Ok();
+            }
+
+            return Forbid();
         }
 
         [HttpPatch("SetNormalizedUserName")]
@@ -66,8 +82,15 @@ namespace UserManaging.API.Controllers
                                                                           string normalizedName,
                                                                           CancellationToken cancellationToken)
         {
-            await _userService.SetNormalizedUserNameAsync(user, normalizedName, cancellationToken);
-            return Ok();
+            var currentUser = await GetCurrentUserAsync();
+
+            if (currentUser.Email.Equals(user.Email) || await IsCurrentUserAdminAsync())
+            {
+                await _userService.SetNormalizedUserNameAsync(user, normalizedName, cancellationToken);
+                return Ok();
+            }
+
+            return Forbid();
         }
 
         [HttpGet("GetUserName")]
@@ -99,5 +122,16 @@ namespace UserManaging.API.Controllers
         {
             return _userService.FindByIdAsync(userId, cancellationToken);
         }
+
+        private async Task<bool> IsCurrentUserAdminAsync()
+        {
+            var currentUser = await GetCurrentUserAsync();
+
+            return await IsUserAdminAsync(currentUser);
+        }
+
+        private async Task<bool> IsUserAdminAsync(UserDTO user) => await _userService.IsInRoleAsync(user, "Admin");
+
+        private async Task<UserDTO> GetCurrentUserAsync() => await HttpUserHelper.GetCurrentUserAsync(_userService, HttpContext);
     }
 }
